backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.18
psycopg[binary]==3.1.18
pydantic-settings==2.3.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.4
rq==1.16.2
httpx==0.27.0
opensearch-py==2.5.0
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
